+++ src/function/configuration/extension-configuration/tests/build-extension-configuration-db.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from "vitest";
 
 import EXTENSIONS from './extensions-state.json'
 import { ExtensionsState } from "../../../extensions/extensions-state";
-import { buildExtensionConfigurationDB } from "../build-extension-configuration-db";
+import { buildExtensionConfigurationDB } from "./build-extension-configuration-db-v1-code.test";
 import { deserializeSimplifiedSerializedExtensionsStateFromExtensions } from "../../../../testing/dump-extensions-state";
 import { addExtensionToExplicityActivatedExtensions } from "../../../../components/ucp-tabs/extension-manager/extensions-state-manipulation";
